Move targetUri extraction utility functions to DnsSupport
